Always include posts by Scala related organization accounts
