+++ src/scala/accounts.ts
@@ -1,0 +1,5 @@
+export const scalaRelatedOrganizationProfiles = [
+    "did:plc:mb6e2ashxeswusv7f7hwusp5", // scala-lang.org https://bsky.social/xrpc/com.atproto.identity.resolveHandle?handle=scala-lang.org
+    "did:plc:5nlshbn5adh3fjwzyz7xjpwl", // Scala Space    https://bsky.social/xrpc/com.atproto.identity.resolveHandle?handle=scalaspace.bsky.social
+    "did:plc:xs35x5l4ogj3g4eymh6ngcsr", // Scala Times    https://bsky.social/xrpc/com.atproto.identity.resolveHandle?handle=scalatimes.com
+]
